MAINT: separated scan utilities
Separated the supporting routines from update_scan into their own file with a header.
